Introduce mainnet block validator builder class
